Finish subjects component
I used as little nesting as possible this time and tried to keep it
clean :)
